pyproject.toml
[project]
name = "siteops_platform"
version = "0.1.0"
description = "Siteops Platform API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "sqlalchemy[asyncio]>=2.0.35",
    "alembic>=1.13",
    "httpx>=0.27",
    "asyncpg>=0.29",
    "redis>=5.0",
    "prometheus-client>=0.21",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27",
    "python-jose[cryptography]>=3.3",
]

[project.optional-dependencies]
dev = [
    "aiosqlite>=0.20",
    "psycopg[binary]>=3.2",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "factory-boy>=3.3",
    "hypothesis>=6.112",
    "respx>=0.21",
    "httpx>=0.27",
    "ruff>=0.6",
    "bandit[toml]>=1.7",
    "pip-audit>=2.7",
    "vulture>=2.11",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
addopts = "-q --strict-markers"
markers = [
    "unit: fast unit tests",
    "integration: DB and external integration",
    "api: HTTP contract tests",
]

[tool.coverage.run]
branch = true
source = ["siteops_platform"]

[tool.coverage.report]
fail_under = 60
show_missing = true

[tool.ruff]
target-version = "py39"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.bandit]
exclude_dirs = ["tests", "alembic"]